refactor(lattices): inline collections/alloc.rs back into collections.rs
Collapse the collections/mod.rs + collections/alloc.rs split back into
a single collections.rs file with inline #[cfg(feature = "alloc")]
gates on VecSet/VecMap and their impls. This produces a much smaller
diff from main (inline cfg additions vs file rename + new file).

PR: #2919
